src/requirements.txt
google-cloud-firestore
flask
pytest
